Ruff (#96)
* First attempt swap adder

* Add tests for swap instertion method

* Remove 313 compaitibility

* Add support for python 313

* Add RoutingBenchmaker

* Linters

* Make Circuit and CircuitLike as benchmark input

* Pylint

* mypy?

* Happify mypy?

* mypy?

* mypy

* Add safe fixes

* Happify ruff

* pydocstyle

* pylint

* Fix mypy issues (numpy bug)

* Happify MYPY

* Update github actions

* Happify ruff

* format

* tst

* Update workflow

* mypy?

* mypy

+++ qgym/envs/routing/routing_state.py
@@ -89,7 +89,7 @@ class RoutingState(State[dict[str, NDArray[np.int_]], int]):
         """An array of 2-tuples of integers, where every tuple represents a, not
         specified, gate acting on the two qubits labeled by the integers in the tuples.
         """
-        self.mapping = np.arange(self.n_qubits)
+        self.mapping = np.arange(self.n_qubits, dtype=np.int_)
         """Array of which each index represents a logical qubit and each value
         represents a physical qubit.
         """
@@ -291,7 +291,7 @@ class RoutingState(State[dict[str, NDArray[np.int_]], int]):
             is_legal_surpass = np.fromiter(
                 iter=starmap(self.is_legal_surpass, interaction_gates_ahead),
                 count=len(interaction_gates_ahead),
-                dtype=np.int8,
+                dtype=np.int_,
             )
             observation["is_legal_surpass"] = is_legal_surpass
